requirements.txt
autopep8==1.6.0
click==8.0.3
dnspython==2.2.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
pymongo==4.0.1
python-dotenv==0.19.2
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
